+++ app/code/ThirdParty/BlogArticle/Block/Adminhtml/Post.php
@@ -4,6 +4,7 @@ namespace ThirdParty\BlogArticle\Block\Adminhtml;
 use Magento\Backend\Block\Template;
 use Magento\Backend\Block\Template\Context;
 use Magento\Framework\Data\Form\FormKey;
+use Magento\Framework\UrlInterface;
 use ThirdParty\BlogArticle\Model\ResourceModel\Post\CollectionFactory;
 
 class Post extends Template
@@ -18,14 +19,23 @@ class Post extends Template
      */
     private $formKey;
 
+    /**
+     * Frontend URL builder (not Admin URL).
+     *
+     * @var UrlInterface
+     */
+    private $frontendUrlBuilder;
+
     public function __construct(
         Context $context,
         CollectionFactory $collectionFactory,
         FormKey $formKey,
+        UrlInterface $frontendUrlBuilder,
         array $data = []
     ) {
         $this->collectionFactory = $collectionFactory;
         $this->formKey = $formKey;
+        $this->frontendUrlBuilder = $frontendUrlBuilder;
         parent::__construct($context, $data);
     }
 
@@ -34,7 +44,9 @@ class Post extends Template
      */
     public function getPosts()
     {
-        return $this->collectionFactory->create();
+        $collection = $this->collectionFactory->create();
+        $collection->setOrder('post_id', 'DESC');
+        return $collection;
     }
 
     /**
@@ -75,5 +87,23 @@ class Post extends Template
                 'form_key' => $this->getFormKey(),
             ]
         );
+    }
+
+    /**
+     * Storefront preview URL for a post.
+     *
+     * @param \ThirdParty\BlogArticle\Model\Post $post
+     * @return string
+     */
+    public function getFrontendViewUrl($post)
+    {
+        $params = ['_nosid' => true];
+        $urlKey = (string) $post->getUrlKey();
+        if ($urlKey !== '') {
+            $params['url_key'] = $urlKey;
+        } else {
+            $params['id'] = (int) $post->getId();
+        }
+        return $this->frontendUrlBuilder->getUrl('blog/post/view', $params);
     }
 }