+++ app/code/ThirdParty/BlogArticle/Controller/Post/View.php
@@ -1,0 +1,63 @@
+<?php
+namespace ThirdParty\BlogArticle\Controller\Post;
+
+use Magento\Framework\App\Action\Action;
+use Magento\Framework\App\Action\Context;
+use Magento\Framework\Controller\Result\ForwardFactory;
+use Magento\Framework\View\Result\PageFactory;
+use ThirdParty\BlogArticle\Model\PostFactory;
+
+class View extends Action
+{
+    /**
+     * @var PageFactory
+     */
+    private $resultPageFactory;
+
+    /**
+     * @var ForwardFactory
+     */
+    private $resultForwardFactory;
+
+    /**
+     * @var PostFactory
+     */
+    private $postFactory;
+
+    public function __construct(
+        Context $context,
+        PageFactory $resultPageFactory,
+        ForwardFactory $resultForwardFactory,
+        PostFactory $postFactory
+    ) {
+        parent::__construct($context);
+        $this->resultPageFactory = $resultPageFactory;
+        $this->resultForwardFactory = $resultForwardFactory;
+        $this->postFactory = $postFactory;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function execute()
+    {
+        $post = $this->postFactory->create();
+        $urlKey = (string) $this->getRequest()->getParam('url_key', '');
+        $id = (int) $this->getRequest()->getParam('id', 0);
+
+        if ($urlKey !== '') {
+            $post->load($urlKey, 'url_key');
+        } elseif ($id) {
+            $post->load($id);
+        }
+
+        if (!$post->getId() || !(int) $post->getIsActive()) {
+            $resultForward = $this->resultForwardFactory->create();
+            return $resultForward->forward('noroute');
+        }
+
+        $resultPage = $this->resultPageFactory->create();
+        $resultPage->getConfig()->getTitle()->set($post->getTitle());
+        return $resultPage;
+    }
+}